% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{Animal}
\alias{Animal}
\title{(Internal) Creates an amial}
\usage{
Animal(x, y, crs)
}
\arguments{
\item{x}{coordinate X}

\item{y}{coordinate Y}

\item{crs}{coordinate reference system}
}
\value{
a new animal
}
\description{
(Internal) Creates an amial
}
\keyword{internal}
