% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{extract_neighbours}
\alias{extract_neighbours}
\title{(Internal) Extracts cell neighbours}
\usage{
extract_neighbours(i, j, nrows, ncols)
}
\arguments{
\item{i}{row id}

\item{j}{column id}

\item{nrows}{total number of rows}

\item{ncols}{total number of columns}
}
\value{
a list of pixel neighbours
}
\description{
(Internal) Extracts cell neighbours
}
\keyword{internal}
