% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{find_connected}
\alias{find_connected}
\title{(Internal) Finds connected pixels}
\usage{
find_connected(r, init_i, init_j)
}
\arguments{
\item{r}{a SpatRaster}

\item{init_i}{ith row to start}

\item{init_j}{jth row to start}
}
\value{
a list
}
\description{
(Internal) Finds connected pixels
}
\keyword{internal}
