% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{is_connected}
\alias{is_connected}
\title{(Internal) Verify if a pixel is connected}
\usage{
is_connected(mat, x, y)
}
\arguments{
\item{mat}{a matrix}

\item{x}{a row id}

\item{y}{a column id}
}
\value{
logical
}
\description{
(Internal) Verify if a pixel is connected
}
\keyword{internal}
