% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{isolate_pixels}
\alias{isolate_pixels}
\title{(Internal) Isolate pixels}
\usage{
isolate_pixels(flowdir, dem, init_i, init_j)
}
\arguments{
\item{flowdir}{a \code{SpatRaser} with the flow directions}

\item{dem}{a \code{SpatRaster} with a digital elevation model}

\item{init_i}{ith row to start}

\item{init_j}{jth column to stat}
}
\value{
a SpatRaster
}
\description{
(Internal) Isolate pixels
}
\keyword{internal}
