% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mec-rootcontact.R
\name{mec_rootcontact}
\alias{mec_rootcontact}
\title{Transmission through root-to-root contact}
\usage{
mec_rootcontact(treecover, aoi, poi, quiet = FALSE)
}
\arguments{
\item{treecover}{A single-band \code{SpatRaster} where 1 represents host
trees, and 0 represents background area}

\item{aoi}{A \code{sf} polygon representing the area of interest. Used to mask
the tree cover}

\item{poi}{A single-point \code{sf} object denoting the point of interest
to run the simulations}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
A \code{SpatRaster}
}
\description{
Simulates diffusive inoculum spread through root contact between
neighbouring trees
}
\details{
This function models inoculum movement due to root-to-root contact based on
tree continuity. It calculates vegetation continuity by counting connected
pixels within a 3×3 window. If continuity exists, a new raster is created with
connected pixels marked as 1. The function then identifies all pixels connected
to the foci within the study area. If no direct pixel-to-pixel contact between
trees exists, the infection risk is considered zero.
}
\examples{
\donttest{
## load packages
library(phytorisk)
library(sf)
library(terra)

## load data
study_area_sf <- st_read(
  system.file("spatial/tejera.geojson", package = "phytorisk"),
  quiet = TRUE
)

poi_sf <- st_read(
  system.file("spatial/poi.geojson", package = "phytorisk"),
  quiet = TRUE
)

trees_sr <- rast(system.file("spatial/trees_light.tiff", package = "phytorisk"))

## simulate mechanism
mec_rootcontact_sr <- mec_rootcontact(trees_sr, study_area_sf, poi_sf)
}
}
\references{
Cardillo, E., Abad, E., Meyer, S., 2021. Iberian oak decline caused by Phytophthora
cinnamomi: A spatiotemporal analysis incorporating the effect of host heterogeneities
at landscape scale. For. Pathol. 51, e12667. \doi{10.1111/efp.12667}

Cardillo, E., Acedo, A., Abad, E., 2018. Topographic effects on dispersal patterns
of Phytophthora cinnamomi at a stand scale in a Spanish heathland. PloS One 13,
e0195060.
}
