% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mec-surfaceflow.R
\name{mec_surfaceflow}
\alias{mec_surfaceflow}
\title{Inoculum spread}
\usage{
mec_surfaceflow(dem, mec_soilwater, poi, buffer = 50, quiet = FALSE)
}
\arguments{
\item{dem}{a single-band \code{SpatRaster} with a digital elevation model}

\item{mec_soilwater}{the result of \link{mec_soilwater}}

\item{poi}{a single-point \code{sf} object denoting the point of interest
to run the simulations}

\item{buffer}{a buffer in meters to extent the spread in every direction}

\item{quiet}{if \code{TRUE}, suppress any message or progress bar}
}
\value{
A \code{SpatRaster}
}
\description{
Simulates the spread of the inoculum in surface water flow
}
\details{
To do...
}
\examples{
## load packages
library(sf)
library(terra)

## load data
poi_sf <- st_read(system.file("spatial/poi.geojson", package = "phytorisk"))
dem_sr <- rast(system.file("spatial/dem_light.tiff", package = "phytorisk"))
trees_sr <- rast(system.file("spatial/trees_light.tiff", package = "phytorisk"))

## simulate mechanism
mec_soilwater_sr <- mec_soilwater(dem_sr, poi_sf)
mec_surface_sr <- mec_surfaceflow(dem_sr, mec_soilwater_sr, poi_sf)
}
