% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{move_towards_food}
\alias{move_towards_food}
\title{(Internal) Move animals towards food}
\usage{
move_towards_food(animal, food_coords, pixel_size, exclusion_area_sf)
}
\arguments{
\item{animal}{location of the animal}

\item{food_coords}{coordinates of food}

\item{pixel_size}{size of the pixel}

\item{exclusion_area_sf}{area of exclusion}
}
\value{
a new animal
}
\description{
Moves animals towards a food source specified in pixels based on randomness
}
\keyword{internal}
