% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-ensemble.R
\name{phytorisk_ensemble_raw}
\alias{phytorisk_ensemble_raw}
\title{Raw Phytophthora ensemble risk}
\usage{
phytorisk_ensemble_raw(
  aoi,
  poi,
  dem,
  treecover,
  weights = "equal",
  th = 100,
  buffer = 50,
  include_zoospread = FALSE,
  append_mec = FALSE,
  ...,
  quiet = FALSE
)
}
\arguments{
\item{aoi}{A \code{sf} polygon representing the area of interest. Used to mask
the tree cover}

\item{poi}{A single-point \code{sf} object denoting the point of interest
to run the simulations}

\item{dem}{A single-band \code{SpatRaster} with a digital elevation model}

\item{treecover}{A single-band \code{SpatRaster} where 1 represents host
trees, and 0 represents background area}

\item{weights}{weights of the ensemble model. The default uses the same
weights for each model. The argument accepts a numeric vector with the
corresponding weights. See \link{phytorisk_ensemble}}

\item{th}{Threshold of flow accumulation to delineate streams}

\item{buffer}{A buffer in meters to extend the spread in every direction}

\item{include_zoospread}{logical. Whether to include the optional module of \link{mec_zoospread}}

\item{append_mec}{Logical. Whether to append to the results of each
individual module to the output SpatRaster}

\item{...}{arguments passed to \link{mec_zoospread}}

\item{quiet}{A logical value. If \code{TRUE}, suppresses any informational messages.
Defaults to \code{FALSE}.}
}
\value{
A \code{SpatRaster}
}
\description{
Calculates Phytophthora cinnamomi's ensemble risk from raw data
}
\examples{
\donttest{
## load packages
library(phytorisk)
library(sf)
library(terra)

## load data
poi_sf <- st_read(
  system.file("spatial/poi.geojson", package = "phytorisk"),
  quiet = TRUE
)
dem_sr <- rast(system.file("spatial/dem_light.tiff", package = "phytorisk"))
trees_sr <- rast(system.file("spatial/trees_light.tiff", package = "phytorisk"))
aoi_sf <- st_read(
  system.file("spatial/tejera.geojson", package = "phytorisk"),
  quiet = TRUE
)

## calculate ensemble risk, returning individual mechanisms
risk_equal_sr <- phytorisk_ensemble_raw(
  aoi = aoi_sf,
  poi = poi_sf,
  dem = dem_sr,
  treecover = trees_sr,
  append_mec = TRUE
)

## visualize results
plot(risk_equal_sr)
}
}
