% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{stay_within_area}
\alias{stay_within_area}
\title{(Internal) Do not leave study area}
\usage{
stay_within_area(animal, aoi)
}
\arguments{
\item{animal}{location of the animal}

\item{aoi}{area of interest (sf)}
}
\value{
animal
}
\description{
(Internal) Do not leave study area
}
\keyword{internal}
