% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{twi}
\alias{twi}
\title{(Internal) Calculates Topographic Water Index}
\usage{
twi(dem)
}
\arguments{
\item{dem}{a SpatRaster}
}
\value{
a SpatRaster
}
\description{
(Internal) Calculates Topographic Water Index
}
\keyword{internal}
