% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mec-rootcontact.R
\name{mec_rootcontact}
\alias{mec_rootcontact}
\title{Root-to-root contact}
\usage{
mec_rootcontact(treecover, aoi, poi, quiet = FALSE)
}
\arguments{
\item{treecover}{a single-band \code{SpatRaster} where 1 represents host
trees, and 0 represents background area}

\item{aoi}{a \code{sf} polygon representing the area of interest}

\item{poi}{a single-point \code{sf} object denotimecng the point of interest
to run the simulations}

\item{quiet}{if \code{TRUE}, suppress any message or progress bar}
}
\value{
A \code{SpatRaster}
}
\description{
Simulates diffusive inoculum spread through root contact
}
\details{
To do...
}
\examples{
\donttest{
## load packages
library(sf)
library(terra)

## load data
study_area_sf <- st_read(system.file("spatial/tejera.geojson", package = "phytorisk"))
poi_sf <- st_read(system.file("spatial/poi.geojson", package = "phytorisk"))
trees_sr <- rast(system.file("spatial/trees_light.tiff", package = "phytorisk"))

## simulate mechanism
mec_rootcontact_sr <- mec_rootcontact(trees_sr, study_area_sf, poi_sf)
}
}
