% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mec-soilwater.R
\name{mec_soilwater}
\alias{mec_soilwater}
\title{Water inoculum movement}
\usage{
mec_soilwater(dem, poi, quiet = FALSE)
}
\arguments{
\item{dem}{a single-band \code{SpatRaster} with a digital elevation model}

\item{poi}{a single-point \code{sf} object denoting the point of interest
to run the simulations}

\item{quiet}{if \code{TRUE}, suppress any message or progress bar}
}
\value{
A \code{SpatRaster}
}
\description{
Simulates inoculum transport via water in the soil matrix
}
\details{
To do...
}
\examples{
## load packages
library(sf)
library(terra)

## load data
dem_sr <- rast(system.file("spatial/dem.tiff", package = "riskPC"))
poi_sf <- st_read(system.file("spatial/poi.geojson", package = "riskPC"))

## simulate mechanism
mec_soilwater_sr <- mec_soilwater(dem_sr, poi_sf)
}
